+++ ibis/backends/base/sql/__init__.py
@@ -1,0 +1,3 @@
+from .client import SQLClient
+
+__all__ = ('SQLClient',)
