+++ ibis/backends/base/sql/alchemy/client.py
@@ -11,8 +11,8 @@ import ibis.expr.datatypes as dt
 import ibis.expr.schema as sch
 import ibis.expr.types as ir
 import ibis.util as util
+from ibis.backends.base.sql import SQLClient
 from ibis.backends.base.sql.compiler import Dialect
-from ibis.client import SQLClient
 
 from .datatypes import to_sqla_type
 from .geospatial import geospatial_supported
